refactor(mcp): split handler.rs to satisfy 500 LOC gate
Extract execute_tool/execute_read_resource dispatch, BM25 indexing, and
hypervector parsing helpers into src/mcp/tools.rs. handler.rs drops from
575 to 243 LOC, tools.rs is 344 LOC. Both under the arch_fitness 500 LOC
gate that was failing CI test and mcp-feature jobs.
